American Flag
Yesterday I was posed with a fun challenge. Build the American flag
using only html & css. I finished the challenge quickly, but wasn’t
happy with how bulky the file was, so I brought jQuery into the
equation and drastically reduced the amount of code needed. I’d say
this is pretty streamlined, but if you have a better method, I’m all
ears!

Looking in the script section, I’ve used two loops to place the stars
and stripes in the flag.